pyproject.toml
[project]
name = "github-language-report"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
